Ultimo commit
Cambios con swal y se agrego validacion para las materias duplicadas

+++ src/components/buttons/DeleteStudentButton.vue
@@ -1,48 +1,41 @@
 <script setup lang="ts">
-import { Dialog, DialogContent, DialogHeader, DialogFooter, DialogTitle, DialogTrigger } from "@/components/ui/dialog";
-import { Button } from '@/components/ui/button';
-import { ref, defineProps, inject } from 'vue';
+import { Button } from "@/components/ui/button";
+import { defineProps, inject } from "vue";
 import axios from "axios";
+import Swal from "sweetalert2";
 
 const props = defineProps<{ id: string }>();
-const dialogOpen = ref(false);
-const reloadData = inject('reloadData') as () => Promise<void>;
-
-const handleDialogClose = () => {
-  dialogOpen.value = false;
-}
+const reloadData = inject("reloadData") as () => Promise<void>;
 
 async function handleConfirmDelete() {
-  try {
-    await axios.delete(`http://localhost:5000/api/student/${props.id}`);
-    await reloadData();
-  } catch (error) {
-    console.error('Error eliminando el estudiante:', error);
-  }
-  dialogOpen.value = false;
+  Swal.fire({
+    title: "Estas seguro?",
+    text: "No podras revertirlo!",
+    icon: "warning",
+    showCancelButton: true,
+    confirmButtonColor: "#3085d6",
+    cancelButtonColor: "#d33",
+    confirmButtonText: "Si, Borralo!",
+  }).then(async (result) => {
+    if (result.isConfirmed) {
+      try {
+        await axios.delete(`http://localhost:5000/api/student/${props.id}`);
+        await reloadData();
+
+        // Mostrar alerta de éxito
+        Swal.fire({
+          title: "Borrado!",
+          text: "Se ha eliminado.",
+          icon: "success",
+        });
+      } catch (error) {
+        console.error("Error eliminando el estudiante:", error);
+      }
+    }
+  });
 }
 </script>
 
 <template>
-  <Dialog v-model:open="dialogOpen">
-    <DialogTrigger as-child>
-      <Button variant="destructive" @click="dialogOpen = true">
-        Eliminar
-      </Button>
-    </DialogTrigger>
-    <DialogContent class="sm:max-w-[425px]">
-      <DialogHeader>
-        <DialogTitle>Confirmar eliminación</DialogTitle>
-      </DialogHeader>
-      <p>¿Estás seguro de que deseas eliminar a este estudiante? Esta acción no se puede deshacer.</p>
-      <DialogFooter>
-        <Button variant="outline" @click="handleDialogClose">Cancelar</Button>
-        <Button variant="destructive" @click="handleConfirmDelete">Eliminar</Button>
-      </DialogFooter>
-    </DialogContent>
-  </Dialog>
+  <Button variant="destructive" @click="handleConfirmDelete"> Eliminar </Button>
 </template>
-
-<style scoped>
-/* Añade estilos adicionales aquí si es necesario */
-</style>
